Improving component visibility notifications
Components were not being properly notified when their contents changed (ie. increased or decreased in count) if they were already visible. This has been resolved by listening for component data changes and re-notifying visible components.
There is a known issue that is when one visible component's data changes, *all* visible components will be notified again. This problem can probably be alleviated once Bento supports proper diffing of component data, but for now it's not a big deal.

Also, moved the visibility listener to its own file as it was bloating the RecyclerViewComponentController quite a bit.

https://reviewboard.yelpcorp.com/r/321714/
https://jira.yelpcorp.com/browse/RX-9918