Wrap the BentoComponent's views in a FrameLayout, so that margins can be applied
The AbsListView.LayoutParam class only contains width and height. When
inflating a view and passing it to a ListView, the layout margins would be
stripped out. This behavior is inconsistant with a RecyclerView. This change
wraps the view created from a modern bento component in a FrameLayout, so that
all ViewGroup.LayoutParams settings can be passed on.

https://reviewboard.yelpcorp.com/r/336509/
https://jira.yelpcorp.com/browse/COREAND-2940