Fixed embedded grid bug.
See ticket for details. My solution is to propagate number of columns and spans
of columns to subcomponents if they are in a ComponentGroup. There is also a
minor speed optimization. Originally, when getting the span size of a
component, we would instantiate a new SpanSizeLookup object each time. This has
some obvious speed concerns as it happens hundreds of times a second. My
solution is to store it in a variable instead of creating a new one each time.

https://reviewboard.yelpcorp.com/r/326410/
https://jira.yelpcorp.com/browse/COREAND-2780