Fixed bento offset bug introduced with recursive columns.
When I added a commit for recursively calculating the number of columns in a
component, I introduced an issue with indexing. The issue is that the start
index of a ComponentGroup is not taken into account when indexing into child
components. Consider the following:

Component A is a ListComponent with 10 children.
Component B is a ListComponent with 2 children.

The existing solution works for components 0-9 in component A. Once we hit an
index of 10, we experience an issue. Index 10 belongs to component B, so we are
currently trying to get the component at index 10 in Component B, which is out
of the range of (0, 2]. The solution for this is to simply subtract the
starting index of the component (Component B in this case) from the index we
are looking for. In this example:

Component A starts at index 0. So, when we get the number of columns at index
1, we calculate index to be (1 - 0) in A, which still gives 1 and works as
expected.
When calculating the number of columns at index 10, we first see that that
component will be in Component B. We then index into B at index (10 - 10) which
is index 0 of B, which is correct in this case.

Note, there are actually two issues that are basically the same. We must make
this change to both the ComponentGroup and the Controller.

https://reviewboard.yelpcorp.com/r/361152/
https://jira.yelpcorp.com/browse/COREAND-2780