Support removal of components in ViewPagerComponent
Support removal of components in ViewPagerComponent.

The ViewPagerComponent was so far used to show a static list of components.
There was no dynamic removal/addition of components.

If we were to remove an added component from the ViewPagerComponent and if that
page is in focus the ViewPagerComponent would have refreshed to show a blank
page.

The following review uses this fix. The ViewPagerComponent and
ViewPagerComponentController was copied to demonstrate this. We will be
removing the copies later.

https://reviewboard.yelpcorp.com/r/369559/

https://reviewboard.yelpcorp.com/r/372402/
https://jira.yelpcorp.com/browse/COREAND-3506