[bento] Stopped resetting the adapter in the ListViewComponentController
Stopped resetting the adapter in the ListViewComponentController to fix a bug
where components calling notifyDataSetChanged were not being updated. Spend
some time debugging this issue and the only thing that fixed it was not
resetting the adapter. By clearing the component group (the data backing the
adapter) I think it should work as expected even without replacing the adapter.

https://reviewboard.yelpcorp.com/r/376790/
https://jira.yelpcorp.com/browse/SEARCHUX-7299