Migrating to AndroidX
Migrating to AndroidX

This change bumps our compile and target SDK versions to 28 and migrates from
support library to the new AndroidX libraries.
Since the consumers of this library are not yet on AndroidX, we need to
dejetify the library before publishing to maven. To achieve this, a new gradle
task has been added as part of the publishing task.

https://reviewboard.yelpcorp.com/r/382179/
https://jira.yelpcorp.com/browse/COREAND-3262