Propagate CarouselComponent visibility to it's children and add a visibility listener to Components
This changes addresses an issue with the CarouselComponent: It would not propagate it's visibility status to it's children.
With this change this is now fixed, and the children components that are currently on screen will get notified if they become visible or invisible.

Also, to avoir people having to subclass component to use visibility listener, adds a set of method to register and unregister listeners.